fix(app navigation): responsiveness and added dropdown in sm
