+++ src/components/accounts/constants/index.js
@@ -69,9 +69,29 @@ const resourceOptions = [
   // { label: 'Specific Content Model', value: 'roles' },
   // { label: 'Specific Content Item', value: 'roles' },
 ];
+
+const developerDocs = [
+  { title: 'Guides and Docs', url: 'https://zesty.org/' },
+  { title: 'Node SDK', url: 'https://github.com/zesty-io/node-sdk' },
+  { title: 'Instance API', url: 'https://instances-api.zesty.org/' },
+  { title: 'Accounts API', url: 'https://accounts-api.zesty.org/' },
+  { title: 'Auth API', url: 'https://auth-api.zesty.org/' },
+  { title: 'Media API', url: 'https://media-api.zesty.org/' },
+  { title: 'Fetch Wrapper', url: 'https://github.com/zesty-io/fetch-wrapper' },
+  { title: 'Status', url: 'https://status.zesty.io/' },
+  { title: 'Parsley', url: 'https://github.com/zesty-io/parsley' },
+];
+const leftNav = [
+  { title: 'Dashboard', url: '/' },
+  { title: 'Instance', url: '/instances/' },
+  { title: 'Teams', url: '/teams/' },
+  { title: 'Marketplace', url: '/marketplace/' },
+];
 export const accounts = {
   eventActionOptions,
   contentTypeOptions,
   methodOptions,
   resourceOptions,
+  developerDocs,
+  leftNav,
 };
