+++ src/components/accounts/ui/menu/DeveloperDocMenu.js
@@ -4,18 +4,7 @@ import MenuItem from '@mui/material/MenuItem';
 import { Box } from '@mui/material';
 import { useRouter } from 'next/router';
 import { NavItem } from 'layouts/Main/components/Topbar/components';
-
-const developerDocs = [
-  { title: 'Guides and Docs', url: 'https://zesty.org/' },
-  { title: 'Node SDK', url: 'https://github.com/zesty-io/node-sdk' },
-  { title: 'Instance API', url: 'https://instances-api.zesty.org/' },
-  { title: 'Accounts API', url: 'https://accounts-api.zesty.org/' },
-  { title: 'Auth API', url: 'https://auth-api.zesty.org/' },
-  { title: 'Media API', url: 'https://media-api.zesty.org/' },
-  { title: 'Fetch Wrapper', url: 'https://github.com/zesty-io/fetch-wrapper' },
-  { title: 'Status', url: 'https://status.zesty.io/' },
-  { title: 'Parsley', url: 'https://github.com/zesty-io/parsley' },
-];
+import { accounts } from 'components/accounts/constants';
 
 export const DeveloperDocMenu = ({}) => {
   const router = useRouter();
@@ -34,7 +23,7 @@ export const DeveloperDocMenu = ({}) => {
       <NavItem
         title={'Documentation'}
         id={'documentation'}
-        items={developerDocs}
+        items={accounts.developerDocs}
         colorInvert={false}
       />
 
@@ -48,7 +37,7 @@ export const DeveloperDocMenu = ({}) => {
         }}
       >
         <Box paddingY={1}>
-          {developerDocs.map((e) => (
+          {accounts.developerDocs.map((e) => (
             <MenuItem title={e.label} onClick={() => handleClose(e.link)}>
               {e.label}
             </MenuItem>
